feat: organising project
