+++ src/datamodel/repository/abstractRepository.ts
@@ -1,4 +1,4 @@
-import { SqlSealDatabase } from "src/database";
+import { SqlSealDatabase } from "src/database/database";
 
 export abstract class Repository {
     constructor(protected readonly db: SqlSealDatabase) { }
