+++ src/renderer/rendererRegistry.ts
